+++ lib/src/models/scan_filter.dart
@@ -1,0 +1,7 @@
+class ScanFilter {
+  final List<String> withServices;
+
+  ScanFilter({
+    this.withServices = const [],
+  });
+}